+++ sdk/src/integrationTest/java/AccountBalanceIntegrationTest.java
@@ -1,6 +1,7 @@
 import com.google.errorprone.annotations.Var;
 import com.hedera.hashgraph.sdk.AccountBalanceQuery;
 import com.hedera.hashgraph.sdk.AccountId;
+import com.hedera.hashgraph.sdk.Client;
 import com.hedera.hashgraph.sdk.Hbar;
 import com.hedera.hashgraph.sdk.PrecheckStatusException;
 import com.hedera.hashgraph.sdk.Status;
@@ -8,6 +9,8 @@ import com.hedera.hashgraph.sdk.TokenCreateTransaction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.Objects;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
@@ -16,6 +19,53 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class AccountBalanceIntegrationTest {
+    @Test
+    @DisplayName("can connect to previewnwet with TLS")
+    void canConnectToPreviewnetWithTLS() {
+        assertDoesNotThrow(() -> {
+            var network = new HashMap<String, AccountId>();
+            network.put("0.previewnet.hedera.com:50212", new AccountId(3));
+            network.put("1.previewnet.hedera.com:50212", new AccountId(4));
+            network.put("2.previewnet.hedera.com:50212", new AccountId(5));
+            network.put("3.previewnet.hedera.com:50212", new AccountId(6));
+            network.put("4.previewnet.hedera.com:50212", new AccountId(7));
+
+            var client = Client.forNetwork(network);
+
+            for (var entry : network.entrySet()) {
+                new AccountBalanceQuery()
+                    .setNodeAccountIds(Collections.singletonList(entry.getValue()))
+                    .setAccountId(entry.getValue())
+                    .execute(client);
+            }
+            client.close();
+        });
+    }
+
+    @Test
+    @DisplayName("can connect to testnet with TLS")
+    void canConnectToTestnetWithTLS() {
+        assertDoesNotThrow(() -> {
+            var network = new HashMap<String, AccountId>();
+            network.put("0.testnet.hedera.com:50212", new AccountId(3));
+            network.put("1.testnet.hedera.com:50212", new AccountId(4));
+            network.put("2.testnet.hedera.com:50212", new AccountId(5));
+            network.put("3.testnet.hedera.com:50212", new AccountId(6));
+            network.put("4.testnet.hedera.com:50212", new AccountId(7));
+
+            var client = Client.forNetwork(network);
+
+            for (var entry : network.entrySet()) {
+                new AccountBalanceQuery()
+                    .setNodeAccountIds(Collections.singletonList(entry.getValue()))
+                    .setAccountId(entry.getValue())
+                    .execute(client);
+            }
+
+            client.close();
+        });
+    }
+
     @Test
     @DisplayName("Can fetch balance for client operator")
     void canFetchBalanceForClientOperator() {