+++ sdk/src/main/java/com/hedera/hashgraph/sdk/AccountCreateTransaction.java
@@ -27,6 +27,7 @@ public final class AccountCreateTransaction extends Transaction<AccountCreateTra
     private Hbar initialBalance = new Hbar(0);
     private boolean receiverSigRequired = false;
     private Duration autoRenewPeriod = DEFAULT_AUTO_RENEW_PERIOD;
+    private int maxAutomaticTokenAssociations = 0;
 
     public AccountCreateTransaction() {
     }
@@ -140,6 +141,16 @@ public final class AccountCreateTransaction extends Transaction<AccountCreateTra
         return this;
     }
 
+    public int getMaxAutomaticTokenAssociations() {
+        return maxAutomaticTokenAssociations;
+    }
+
+    public AccountCreateTransaction setMaxAutomaticTokenAssociations(int amount) {
+        requireNotFrozen();
+        maxAutomaticTokenAssociations = amount;
+        return this;
+    }
+
     public String getAccountMemo() {
         return accountMemo;
     }
@@ -152,18 +163,20 @@ public final class AccountCreateTransaction extends Transaction<AccountCreateTra
     }
 
     CryptoCreateTransactionBody.Builder build() {
-        var builder = CryptoCreateTransactionBody.newBuilder();
+        var builder = CryptoCreateTransactionBody.newBuilder()
+            .setInitialBalance(initialBalance.toTinybars())
+            .setReceiverSigRequired(receiverSigRequired)
+            .setAutoRenewPeriod(DurationConverter.toProtobuf(autoRenewPeriod))
+            .setMemo(accountMemo)
+            .setMaxAutomaticTokenAssociations(maxAutomaticTokenAssociations);
 
         if (proxyAccountId != null) {
             builder.setProxyAccountID(proxyAccountId.toProtobuf());
         }
+
         if (key != null) {
             builder.setKey(key.toProtobufKey());
         }
-        builder.setInitialBalance(initialBalance.toTinybars());
-        builder.setReceiverSigRequired(receiverSigRequired);
-        builder.setAutoRenewPeriod(DurationConverter.toProtobuf(autoRenewPeriod));
-        builder.setMemo(accountMemo);
 
         return builder;
     }
@@ -190,6 +203,7 @@ public final class AccountCreateTransaction extends Transaction<AccountCreateTra
         initialBalance = Hbar.fromTinybars(body.getInitialBalance());
         accountMemo = body.getMemo();
         receiverSigRequired = body.getReceiverSigRequired();
+        maxAutomaticTokenAssociations = body.getMaxAutomaticTokenAssociations();
     }
 
     @Override