+++ sdk/src/main/java/com/hedera/hashgraph/sdk/Client.java
@@ -82,11 +82,7 @@ public final class Client implements AutoCloseable, WithPing, WithPingAll {
     }
 
     public List<String> getMirrorNetwork() {
-        if (mirrorNetwork != null) {
-            return mirrorNetwork.addresses;
-        } else {
-            return new ArrayList<>();
-        }
+        return mirrorNetwork.getNetwork();
     }
 
     /**
@@ -105,7 +101,7 @@ public final class Client implements AutoCloseable, WithPing, WithPingAll {
     public static Client forNetwork(Map<String, AccountId> networkMap) {
         var executor = createExecutor();
         var network = Network.forNetwork(executor, networkMap);
-        var mirrorNetwork = MirrorNetwork.forMainnet(executor);
+        var mirrorNetwork = MirrorNetwork.forNetwork(executor, new ArrayList<>());
 
         return new Client(executor, network, mirrorNetwork);
     }
@@ -147,7 +143,7 @@ public final class Client implements AutoCloseable, WithPing, WithPingAll {
     public static Client forTestnet() {
         var executor = createExecutor();
         var network = Network.forTestnet(executor);
-        var mirrorNetwork = MirrorNetwork.forMainnet(executor);
+        var mirrorNetwork = MirrorNetwork.forTestnet(executor);
 
         return new Client(executor, network, mirrorNetwork);
     }
@@ -155,7 +151,7 @@ public final class Client implements AutoCloseable, WithPing, WithPingAll {
     public static Client forPreviewnet() {
         var executor = createExecutor();
         var network = Network.forPreviewnet(executor);
-        var mirrorNetwork = MirrorNetwork.forMainnet(executor);
+        var mirrorNetwork = MirrorNetwork.forPreviewnet(executor);
 
         return new Client(executor, network, mirrorNetwork);
     }