+++ CHANGELOG.md
@@ -6,10 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+### v2.0.13
+
+### Added
+
+ * `Account[Create|Update]Transaction.[get|set]MaxAutomaticTokenAssociations`
+ * `TokenAssociation` and `TransactionRecord.automaticTokenAssociations`
+ * `AccountInfo.maxAutomaticTokenAssociations`
+ * `TokenRelationship.automaticAssociation`
+ * New status codes
+
+### Fixed
+
+ * TLS connector failing when the networks address book did not have cert hashes
+
 ### v2.0.12
 
 ### Added
 
+ * Support for TLS connections with Hedera Services nodes when network addresses end in `50212` or `443`
  * Added `FeeAssessmentMethod`.
  * Added `[get|set]AssessmentMethod()` to `CustomFractionalFee`
  * Added `CustomRoyaltyFee`